Miscellaneous code cleanups and fixes: - Updated HTTP GetTile request to replace scale (double) with scale index (INT32). - Commented out unused OGC framework parameters (due to Insure++ errors on Linux). - Got rid of some compiling warnings in Server Tile Service.
git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@1067 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Server/src/Services/Tile/ServerTileService.cpp
@@ -119,7 +119,7 @@ MgByteReader* MgServerTileService::GetTile(
                 cachedMap = new MgMemoryStreamHelper();
                 Ptr<MgStream> stream = new MgStream(cachedMap);
                 map->Serialize(stream);
-                if (sm_mapCache.size() > sm_mapCacheSize)
+                if ((INT32)sm_mapCache.size() > sm_mapCacheSize)
                 {
                     ClearMapCache(L"");
                 }
